python/agents/genmedia-for-commerce/pyproject.toml
[project]
name = "genmedia-for-commerce"
version = "0.1.0"
description = "GenMedia for Commerce - ADK Agent with MCP tools for retail media generation"
requires-python = ">=3.11,<3.14"
dependencies = [
    # ADK & Agent infrastructure
    "google-adk>=1.15.0,<2.0.0",
    "opentelemetry-instrumentation-google-genai>=0.1.0,<1.0.0",
    "gcsfs>=2024.11.0",
    "google-cloud-logging>=3.12.0,<4.0.0",
    "google-cloud-aiplatform[agent-engines,evaluation]>=1.130.0",
    "asyncpg>=0.30.0,<1.0.0",
    # Web framework
    "fastapi>=0.115.8,<1.0.0",
    "uvicorn[standard]>=0.34.0",
    "python-multipart>=0.0.20",
    "python-dotenv>=1.0.1",
    # Google AI
    "google-genai>=1.62.0",
    "google-auth>=2.47.0",
    # Image/Video processing
    "scikit-image>=0.25.2",
    "opencv-python-headless>=4.12.0.88",
    "imageio>=2.37.0",
    "imageio-ffmpeg>=0.6.0",
    "moviepy>=2.2.1",
    "rembg>=2.0.57",
    "google-cloud-vision>=3.9.0",
    "google-cloud-storage>=2.19.0",
    # ML / Eval
    "insightface>=0.7.3",
    "onnxruntime>=1.17.0",
    # Utilities
    "jinja2>=3.1.6",
    "matplotlib>=3.10.7",
    "scikit-learn>=1.7.2",
    "tabulate>=0.9.0",
    "seaborn",
    "pandas",
    "db-dtypes",
    # GCP services
    "google-cloud-bigquery-storage>=2.34.0",
    # MCP
    "mcp[cli]>=1.0.0",
    # Catalogue search (in-memory)
    "pyarrow",
    "protobuf>=6.31.1,<7.0.0",
]

[dependency-groups]
dev = [
    "pytest>=8.3.4,<9.0.0",
    "pytest-asyncio>=0.23.8,<1.0.0",
    "nest-asyncio>=1.6.0,<2.0.0",
]

[project.optional-dependencies]
jupyter = [
    "ipykernel",
    "ipywidgets",
    "jupyter",
]
eval = [
    # Note: google-adk[eval] requires google-cloud-aiplatform[evaluation]
    # which caps scikit-learn<=1.5.2, conflicting with scikit-learn>=1.7.2
    # needed by this project's image processing code.
    # Install eval deps separately when needed: pip install google-adk[eval]
]
lint = [
    "ruff>=0.4.6,<1.0.0",
    "codespell>=2.2.0,<3.0.0",
]
test = [
    "pytest>=8.0.0",
    "pytest-cov>=4.0.0",
    "pytest-asyncio>=0.23.8,<1.0.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["genmedia4commerce"]

[tool.pytest.ini_options]
pythonpath = [".", "genmedia4commerce"]
asyncio_default_fixture_loop_scope = "function"

[tool.ruff]
line-length = 88
target-version = "py310"

[tool.ruff.lint]
select = ["E", "F", "W", "I", "B", "UP", "RUF"]
ignore = ["E501", "C901", "B006"]

[tool.ruff.lint.isort]
known-first-party = ["genmedia4commerce", "agents", "mcp_server", "workflows"]

[tool.codespell]
ignore-words-list = "rouge"
skip = "uv.lock,.venv,src/frontend,**/*.ipynb,**/package-lock.json,**/node_modules"

[tool.agent-starter-pack]
name = "genmedia-for-commerce"
description = "GenMedia retail media generation agent with MCP tools"
base_template = "adk"
agent_directory = "genmedia4commerce"
